refactor: Move course card components to the courses list screen folder
