+++ src/screens/courses-list/components/CourseCard.tsx
