+++ src/screens/courses-list/components/CourseCardSkeleton.tsx
@@ -1,4 +1,4 @@
-import { Skeleton } from "../ui/skeleton";
+import { Skeleton } from "@/components/ui/skeleton";
 
 export const CourseCardSkeleton = () => {
   return (
