+++ src/screens/courses-list/dialogs/create-course/CreateCourseDialog.tsx
@@ -1,4 +1,3 @@
-import { ButtonIconContainer } from "@/components/CourseCard/ButtonIconContainer";
 import {
   Dialog,
   DialogContent,
@@ -10,6 +9,7 @@ import {
 import { Plus } from "lucide-react";
 import { useState } from "react";
 
+import { ButtonIconContainer } from "../../components/ButtonIconContainer";
 import { CreateCourseForm } from "./CreateCourseForm";
 
 export const CreateCourseDialog = () => {
